+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - SignedIndexArray.equals()
 - class IndexArray<EL>, same as SignedIndexArray but for non-negative indices.
+- Added KVComponent interface for Map1, Map2, Map3, IndexArray and SignedIndexArray.
 
 ## [1.14.0] - 2025-10-19
 ### Added